+++ modules/vehicle/vehicle_can_codec.h
@@ -1,0 +1,30 @@
+#pragma once
+
+#include "modules/can/can_frame.h"
+#include "modules/vehicle/VehicleState.h"
+
+#include <cstdint>
+
+namespace cockpit {
+namespace vehicle {
+
+enum class VehicleCanDecodeStatus {
+  kDecoded,
+  kIgnored,
+  kInvalid,
+};
+
+class VehicleCanCodec {
+ public:
+  static constexpr std::uint32_t kStateFrameId = 0x123;
+  static constexpr std::uint8_t kStateFrameLength = 5;
+
+  static can::CanFrame Encode(const VehicleState& state);
+  static VehicleCanDecodeStatus Decode(const can::CanFrame& frame, VehicleState* state);
+
+ private:
+  static std::uint16_t EncodeSpeed(double speed_kph);
+};
+
+}  // namespace vehicle
+}  // namespace cockpit