+++ tools/can-simulator/main.cc
@@ -1,7 +1,9 @@
-#include "common/can/can_frame.h"
-#include "common/can/socket_can.h"
-#include "common/logging/Logger.h"
-#include "common/runtime/ServiceRuntime.h"
+#include "modules/can/can_frame.h"
+#include "modules/vehicle/VehicleState.h"
+#include "modules/vehicle/vehicle_can_codec.h"
+#include "drivers/socketcan/socket_can.h"
+#include "core/logging/Logger.h"
+#include "core/runtime/ServiceRuntime.h"
 
 #include <array>
 #include <chrono>
@@ -12,11 +14,8 @@
 namespace {
 
 cockpit::can::CanFrame MakeMockCanFrame(int sequence) {
-  std::array<std::uint8_t, cockpit::can::CanFrame::kMaxDataLength> data{};
-  for (std::size_t i = 0; i < data.size(); ++i) {
-    data[i] = static_cast<std::uint8_t>((sequence + static_cast<int>(i)) & 0xFF);
-  }
-  return cockpit::can::CanFrame(0x123, data, static_cast<std::uint8_t>(data.size()));
+  return cockpit::vehicle::VehicleCanCodec::Encode(
+      cockpit::vehicle::MakeMockVehicleState(sequence));
 }
 
 }  // namespace