+++ tools/topic/main.cc
@@ -1,4 +1,4 @@
-#include "common/config/Config.h"
+#include "core/config/Config.h"
 #include "topic_command_line.h"
 #include "topic_echo.h"
 #include "topic_hz.h"